snoqtt-alpine-pcap/require/requirements.txt
paho-mqtt
flask
requests
tinydb
dpkt
snortunsock